backend/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
celery>=5.3.0
redis>=5.0.0
psycopg2-binary>=2.9.0
cerebras-cloud-sdk>=1.0.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
dj-database-url>=2.1.0